Assignment3/requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
googlesearch-python>=1.2.3
beautifulsoup4>=4.9.3
google-genai
rich
markdown2
